[ion] millis() and micros() returns uint32_t and are defined in their own header

+++ ion/src/simulator/init.cpp
@@ -119,14 +119,14 @@ void Ion::msleep(long ms) {
 
 static auto start = std::chrono::high_resolution_clock::now();
 
-long Ion::millis() {
+uint32_t Ion::millis() {
   sDisplay->redraw();
   Fl::wait(0);
   auto elapsed = std::chrono::high_resolution_clock::now() - start;
   return std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count();
 }
 
-long Ion::micros() {
+uint32_t Ion::micros() {
   sDisplay->redraw();
   Fl::wait(0);
   auto elapsed = std::chrono::high_resolution_clock::now() - start;
